+++ R/trakt.user.friendfollow.R
@@ -3,7 +3,9 @@
 #' \code{trakt.user.friends} pulls a user's friends, the two-way relationship
 #' of both following and being followed by a user.
 #' Since no OAuth2 methods are supported yet, the specified user mustn't be private.
-#' @param user Target user. Defaults to \code{getOption("trakt.username")}
+#' @param user Target user. Defaults to \code{getOption("trakt.username")}. If multiple users
+#' are specified, the results will be \code{rbind}ed together and a \code{source_user} variable is
+#' appended to indicated which user belongs to wich input user.
 #' @param extended Either \code{min} for standard info, \code{full} for details or \code{full,images}
 #' for additional avatar URLs.
 #' @return A \code{data.frame} containing user information.
@@ -19,7 +21,14 @@ trakt.user.friends <- function(user = getOption("trakt.username"), extended = "m
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }
-
+  if (length(user) > 1){
+    response <- plyr::ldply(user, function(user){
+      response <- trakt.user.friends(user = user, extended = extended)
+      response$source_user <- user
+      return(response)
+    })
+    return(response)
+  }
   # Construct URL, make API call
   url      <- build_trakt_url("users", user, "friends", extended = extended)
   response <- trakt.api.call(url = url)
@@ -40,7 +49,9 @@ trakt.user.friends <- function(user = getOption("trakt.username"), extended = "m
 #'
 #' \code{trakt.user.followers} pulls a user's followers
 #' Since no OAuth2 methods are supported yet, the specified user mustn't be private.
-#' @param user Target user. Defaults to \code{getOption("trakt.username")}
+#' @param user Target user. Defaults to \code{getOption("trakt.username")}. If multiple users
+#' are specified, the results will be \code{rbind}ed together and a \code{source_user} variable is
+#' appended to indicated which user belongs to wich input user.
 #' @param extended Either \code{min} for standard info, \code{full} for details or \code{full,images}
 #' for additional avatar URLs.
 #' @return A \code{data.frame} containing user information.
@@ -56,7 +67,14 @@ trakt.user.followers <- function(user = getOption("trakt.username"), extended = 
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }
-
+  if (length(user) > 1){
+    response <- plyr::ldply(user, function(user){
+      response <- trakt.user.followers(user = user, extended = extended)
+      response$source_user <- user
+      return(response)
+    })
+    return(response)
+  }
   # Construct URL, make API call
   url      <- build_trakt_url("users", user, "followers", extended = extended)
   response <- trakt.api.call(url = url)
@@ -77,7 +95,9 @@ trakt.user.followers <- function(user = getOption("trakt.username"), extended = 
 #'
 #' \code{trakt.user.following} pulls a user's followings.
 #' Since no OAuth2 methods are supported yet, the specified user mustn't be private.
-#' @param user Target user. Defaults to \code{getOption("trakt.username")}
+#' @param user Target user. Defaults to \code{getOption("trakt.username")}. If multiple users
+#' are specified, the results will be \code{rbind}ed together and a \code{source_user} variable is
+#' appended to indicated which user belongs to wich input user.
 #' @param extended Either \code{min} for standard info, \code{full} for details or \code{full,images}
 #' for additional avatar URLs.
 #' @return A \code{data.frame} containing user information.
@@ -93,7 +113,14 @@ trakt.user.following <- function(user = getOption("trakt.username"), extended = 
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }
-
+  if (length(user) > 1){
+    response <- plyr::ldply(user, function(user){
+      response <- trakt.user.following(user = user, extended = extended)
+      response$source_user <- user
+      return(response)
+    })
+    return(response)
+  }
   # Construct URL, make API call
   url      <- build_trakt_url("users", user, "following", extended = extended)
   response <- trakt.api.call(url = url)