+++ NEWS.md
@@ -8,6 +8,8 @@
     * Requires `extended` to be more than `min` since the required `aired_episodes` field
     is only present with higher levels of detail.
 * **Add** `extended` param to `trakt.user.f*`
+* **Add** multiple user input support for `trakt.user.f*` functions.
+    * Results will be `rbind`ed together and a `source_user` column is appended.
 * **Fix** `trakt.user.f*` now return `NULL` when the user is private instead of failing.
 
 ### v0.11.3